BACKEND/requirements.txt
deep-translator
bson
fastapi
typing
starlette
passlib
python-jose
python-dotenv
pydantic
pydantic[email]
numpy
sentence-transformers
langchain
ollama
asyncio
Python-IO
pillow
bcrypt
pymongo
email-to
uvicorn
langdetect

